+++ pages/genFinishedProduct.php
@@ -363,7 +363,7 @@ $.ajax({
 					return;
 				}
 				
-				$cats_q = mysqli_query($conn, "SELECT name FROM IFRACategories ORDER BY id ASC");
+				$cats_q = mysqli_query($conn, "SELECT name,description FROM IFRACategories ORDER BY id ASC");
 				while($cats_res = mysqli_fetch_array($cats_q)){
 					$cats[] = $cats_res;
 				}
@@ -419,7 +419,7 @@ $.ajax({
     <td>Category Class:</td>
     <td><select name="defCatClass" id="defCatClass" class="form-control selectpicker" data-live-search="true">
 		<?php foreach ($cats as $IFRACategories) {?>
-				<option value="cat<?php echo $IFRACategories['name'];?>" <?php echo ($settings['defCatClass']=='cat'.$IFRACategories['name'])?"selected=\"selected\"":""; ?>><?php echo 'Cat '.$IFRACategories['name'];?></option>
+				<option value="cat<?php echo $IFRACategories['name'];?>" <?php echo ($settings['defCatClass']=='cat'.$IFRACategories['name'])?"selected=\"selected\"":""; ?>><?php echo 'Cat '.$IFRACategories['name'].' - '.$IFRACategories['description'];?></option>
 		  <?php	}	?>
             </select></td>
     <td>&nbsp;</td>