+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New:
 ----------------------
 - Make formula page improvements
+- Formula is now automatically marked as made and under evaluation status when making is complete